[Feat] 평균 퀴즈 풀이 점수 조회 컨트롤러 분리
