+++ src/main/java/com/mtvs/sejong/question/domain/service/QuestionService.java
@@ -23,7 +23,7 @@ public class QuestionService {
     public List<RecommendRequestDTO.QuestionDTO> getRecommendedQuestions() {
 
         // user weak question type 확인
-        String weakQuestionType = "띄어쓰기";
+        String weakQuestionType = "문해력";
 
         // weak_question_type으로 질문 검색
         List<Question> questionList = questionRepository.findByQuestionType(weakQuestionType);
